dropzone click fix; proper cleanup implemented;
